requirements.txt
# ── Framework ─────────────────────────────────────────────────────────────
django==5.1.*
django-ninja==1.6.0
django-cors-headers==4.6.*
django-environ==0.11.*

# ── Database ──────────────────────────────────────────────────────────────
psycopg[binary]==3.2.*

# ── Cache / Queue ─────────────────────────────────────────────────────────
redis==5.2.*
django-redis==5.4.*
celery==5.4.*

# ── AI / RAG ──────────────────────────────────────────────────────────────
langgraph==0.2.*
langchain==0.3.*
langchain-openai==0.2.*
langchain-qdrant==0.2.*
langchain-community==0.3.*
qdrant-client==1.12.*
openai==1.57.*

# ── HTTP ──────────────────────────────────────────────────────────────────
httpx==0.28.*

# ── Object storage ────────────────────────────────────────────────────────
minio==7.2.*

# ── Serving ───────────────────────────────────────────────────────────────
gunicorn==23.*
uvicorn[standard]==0.32.*

# ── Dev / testing ─────────────────────────────────────────────────────────
pytest==8.*
pytest-django==4.*
pytest-asyncio==0.24.*
ruff==0.8.*
